alterando tipo de variável

+++ PrimeiraSemanaJava/CalcularSalario/src/CalcularSalario.java
@@ -4,8 +4,8 @@ public class CalcularSalario {
 
         float salario = 1500;
         float totalVendas = 300;
-        double comissao = (totalVendas/100)*10;
-        double salarioTotal = salario + comissao;
+        float comissao = (totalVendas/100)*10;
+        float salarioTotal = salario + comissao;
 
         System.out.printf("Seu salario somado com a comissao eh: R$ %.2f", salarioTotal);
